Ensure endianness is set correctly with GCC
GCC has standard predefined macros for byte order, documented at https://gcc.gnu.org/onlinedocs/cpp/Standard-Predefined-Macros.html. None of them are being used here, and there are no tests for 64-bit architectures like x86_64, so the result is failure to detect endianness when building with GCC on x86_64, which is a fairly standard setup.

This is not necessarily the cleanest-possible fix, especially since it does not clean up below comments about undocumented macros that GCC may or may not define, which do not seem to be accurate. The macro magic here could certainly stand some further cleanup.